+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Environment variables may now be used in YAML configuration (#145).
+
 ## [4.2.0] - 2023-01-13
 
 ### Added